unescape ansi codes, cleanup
