+++ frontend/src/components/status/ErrorMessage.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement } from "react";
-import { useSyntaxHighlighter } from "../../hooks/use-syntax-highlighter";
+import { TextUtils } from "../../utils/text-utils";
 
 export type TransactionErrorMessageProps = {
   errorMessage: string;
@@ -8,14 +8,14 @@ export type TransactionErrorMessageProps = {
 export function TransactionErrorMessage({
   errorMessage,
 }: TransactionErrorMessageProps): ReactElement {
-  const { highlightCadenceSyntax } = useSyntaxHighlighter();
-  const highlightedError = highlightCadenceSyntax(errorMessage);
-
   return (
     <pre
       style={{ height: 500, overflow: "scroll" }}
       dangerouslySetInnerHTML={{
-        __html: highlightedError.replaceAll("\\n", "<br/>"),
+        __html: TextUtils.highlightCadenceSyntax(errorMessage).replaceAll(
+          "\\n",
+          "<br/>"
+        ),
       }}
     />
   );
