--- frontend/src/pages/accounts/details/Storage.module.scss
@@ -1,181 +1,0 @@
-@import 'styles/spacings';
-@import 'styles/colors';
-@import 'styles/typography';
-
-.root {
-  display: flex;
-
-  .storageContent {
-    flex: 3;
-    padding-left: $spacing-l;
-    border-left: 1px solid $yellow;
-
-    pre {
-      color: #f8f8f2;
-      background: none;
-      font-family: Consolas, Monaco, 'Andale Mono', 'Ubuntu Mono', monospace;
-      font-size: $font-size-normal;
-      text-align: left;
-      white-space: pre-wrap;
-      word-spacing: normal;
-      word-break: normal;
-      word-wrap: normal;
-      line-height: 1.5;
-
-      -moz-tab-size: 4;
-      -o-tab-size: 4;
-      tab-size: 4;
-
-      -webkit-hyphens: none;
-      -moz-hyphens: none;
-      -ms-hyphens: none;
-      hyphens: none;
-
-      [class="token comment"],
-      [class="token prolog"],
-      [class="token doctype"],
-      [class="token cdata"],
-      .token.comment,
-      .token.prolog,
-      .token.doctype,
-      .token.cdata {
-        color: #8292a2;
-      }
-
-      [class="token punctuation"],
-      .token.punctuation {
-        //color: #f8f8f2;
-        color: #8292a2;
-      }
-
-      [class="token namespace"],
-      .token.namespace {
-        opacity: .7;
-      }
-
-      [class="token property"],
-      [class="token tag"],
-      [class="token constant"],
-      [class="token symbol"],
-      [class="token deleted"],
-      .token.property,
-      .token.tag,
-      .token.constant,
-      .token.symbol,
-      .token.deleted {
-        color: $yellow;
-
-      }
-
-      [class="token cadence"],
-      .token.cadence {
-        //color: pink;
-        color: #3188CA;
-      }
-
-      [class="token boolean"],
-      [class="token number"],
-      .token.boolean,
-      .token.number {
-        color: $yellow;
-      }
-
-      [class="token selector"],
-      [class="token attr-name"],
-      [class="token string"],
-      [class="token char"],
-      [class="token builtint"],
-      [class="token inserted"],
-      .token.selector,
-      .token.attr-name,
-      .token.string,
-      .token.char,
-      .token.builtin,
-      .token.inserted {
-        color: $light-grey;
-      }
-
-      [class="token operator"],
-      [class="token entity"],
-      [class="token url"],
-      [class="token string"],
-      [class="token variable"],
-      .token.operator,
-      .token.entity,
-      .token.url,
-      .language-css .token.string,
-      .style .token.string,
-      .token.variable {
-        color: #f8f8f2;
-      }
-
-      [class="token atrule"],
-      [class="token value"],
-      [class="token function"],
-      [class="token class-name"],
-      .token.atrule,
-      .token.attr-value,
-      .token.function,
-      .token.class-name {
-        color: hsl(75, 70%, 60%);
-      }
-
-      [class="token keyword"],
-      .token.keyword {
-        color: $blue;
-      }
-
-      [class="token regex"],
-      [class="token important"],
-      .token.regex,
-      .token.important {
-        color: #fd971f;
-      }
-
-      [class="token important"],
-      [class="token bold"],
-      .token.important,
-      .token.bold {
-        font-weight: bold;
-      }
-
-      [class="token italic"],
-      .token.italic {
-        font-style: italic;
-      }
-
-      [class="token entity"],
-      .token.entity {
-        cursor: help;
-      }
-    }
-  }
-
-  .listItems {
-    padding-top: $spacing-base * 2;
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-
-    .tab {
-      border-bottom-right-radius: 0;
-      border-top-right-radius: 0;
-      margin-bottom: $spacing-base;
-      cursor: pointer;
-      box-shadow: inset -7px 0 9px -7px rgba(0,0,0,0.4);
-
-      &.active {
-        //background-color: red;
-        box-shadow: none;
-        border: 1px solid $yellow;
-        border-right: none;
-        z-index: 100;
-        margin-right: -1px;
-      }
-    }
-
-    //border: 1px solid red;
-
-  }
-}
-
